requirements.txt
python-telegram-bot~=22.3
beautifulsoup4~=4.9.3
requests~=2.25.1
python-dotenv~=1.0.0